Moved RecordPutTransformer to cdap-etl-common
