+++ cdap-app-templates/cdap-etl/cdap-etl-common/src/test/java/co/cask/cdap/etl/common/RecordPutTransformerTest.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.template.etl.common;
+package co.cask.cdap.etl.common;
 
 import co.cask.cdap.api.common.Bytes;
 import co.cask.cdap.api.data.format.StructuredRecord;
