+++ cdap-app-templates/cdap-etl/cdap-etl-lib/src/main/java/co/cask/cdap/template/etl/common/Properties.java
@@ -16,6 +16,8 @@
 
 package co.cask.cdap.template.etl.common;
 
+import co.cask.cdap.etl.common.RecordPutTransformer;
+
 /**
  * Class to define property names for source and sinks
  */
@@ -122,7 +124,7 @@ public final class Properties {
     public static final String PROPERTY_SCHEMA = co.cask.cdap.api.dataset.table.Table.PROPERTY_SCHEMA;
     public static final String PROPERTY_SCHEMA_ROW_FIELD =
       co.cask.cdap.api.dataset.table.Table.PROPERTY_SCHEMA_ROW_FIELD;
-    public static final String CASE_SENSITIVE_ROW_FIELD = "case.sensitive.row.field";
+    public static final String CASE_SENSITIVE_ROW_FIELD = RecordPutTransformer.CASE_SENSITIVE_ROW_FIELD;
   }
 
   /**
